fix reading files
the reading algorithm now reads files up to the memory buffer size,
rather than leaving some of it empty if it's not divisible by the block
size.

since all the files are sorted by the x value, the first and last x
values in the read block are checked if they are bigger and smaller than
the maximum and minimum x value respectively, in which case the block is
discarded.